+++ gherkin/go/cmd/main.go
@@ -9,9 +9,9 @@ import (
 	b64 "encoding/base64"
 	"flag"
 	"fmt"
-	"github.com/cucumber/gherkin-go/v9"
-	messages "github.com/cucumber/messages-go/v9"
-	fio "github.com/cucumber/messages-go/v9/io"
+	"github.com/cucumber/gherkin-go/v10"
+	messages "github.com/cucumber/messages-go/v10"
+	fio "github.com/cucumber/messages-go/v10/io"
 	gio "github.com/gogo/protobuf/io"
 	"os"
 )