+++ json-formatter/go/cmd/main.go
@@ -3,8 +3,8 @@ package main
 import (
 	"flag"
 	"fmt"
-	jsonFormatter "github.com/cucumber/json-formatter-go/v2"
-	fio "github.com/cucumber/messages-go/v9/io"
+	jsonFormatter "github.com/cucumber/json-formatter-go/v3"
+	fio "github.com/cucumber/messages-go/v10/io"
 	gio "github.com/gogo/protobuf/io"
 	"io"
 	"log"